moving test address to fixtures folder
